Hooked up GUI card buttons.
User can now select which maneuver they wish to perform by clicking the
appropriate button on the HUD. The card will be placed on the board, and
the pane will fly towards the destination.

With this in place a lot of the testing functionality has been removed
(eg. flying to the big white spheres).